web/requirements.txt
awscli==1.16.28
botocore==1.12.18
certifi==2019.9.11
Click==7.0
colorama==0.3.9
docutils==0.14
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.3
lxml==4.4.2
MarkupSafe==1.1.1
pyasn1==0.4.4
python-dateutil==2.7.3
python-docx==0.8.10
PyYAML==3.13
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
Werkzeug==0.16.0
